Xamarin-2562: Don't ignore padding on controls
ScrollableControl needs to use Padding for calculating DisplayRectangle
because DockPadding is calculated lazily.

When positioning text on button we have to use the internal property
PaddingClientRectangle which takes padding into account.

When drawing a CheckBox we need to consider padding when calculating
the position of the checkbox glyph and of the text.

Change-Id: I5301e36741648696f69919e624936bf37c0e0a73