Add gtk#
We also have to build GTK# so that we can use the GTK# OpenFileDialog.
Although I don't expect that we'll have to provide any patches for
GTK# we still have to compile our own version so that the assemblies
end up in our version of the Mono GAC.

Copying the GTK# assemblies from the installed package doesn't work
since those are compiled with 4.0 framework (at least on Oneiric),
while for FieldWorks we still use the 2.0 framework. Trying to open
the FileChooserDialog in that case results in a "Method not found:
System.Type.op_inequality" exception.

Creating GTK# in our custom location is in line with the docs at
http://www.mono-project.com/Parallel_Mono_Environments.

Reviewed by: Neil Mayhew

Change-Id: Ic1f50fd37733f194eeb1134c52378efce9efee3c