Fix mono-basic build problem in oneiric
Oneiric uses a newer version of dh_clideps (part of cli-common-dev)
which makes it an error to have an unresolved moduleref in an assembly.
Instead, it provides an option to exclude specific module names and
predefines a default list of module names that correspond to imported
Windows-only dlls that need to be treated as errors on Linux.

Unfortunately, it doesn't include ole32.dll in that list, so we need to
list it explicitly. Although the option isn't supported in
cli-common-dev < 0.8 (ie pre-oneiric) fortunately it's ignored with a
warning.

I reported the non-exclusion of ole32 as Debian bug #657564

Change-Id: Id84197e8e1e35b8d68c9f257b1ed3bc0d2e2ea71