Enable building gtk-sharp with glib >= 2.32
The threads API changed in glib 2.32. Later versions of gtk-sharp have
adjusted to this, but 2.12.10 doesn't have these changes. This fix
checks the version of glib during configuration and includes the
deprecated version of the gthreads header if necessary.

glib2.32 is in Ubuntu Precise and Debian Unstable, so this change is
necessary for building gtk-sharp 2.12.10 on either of those systems.

Change-Id: I576364d72dd723ab9d25108e6dc152a56a0447cb