Xamarin-5278: Fix Binder.SelectMethod
This patch fixes the problem that reflection doesn't find the
constructor of a class if one of the arguments is a COM object
implemented in unmanaged code. In that case the proxy can't
be converted to a managed object and we didn't check if the
COM object implements the desired interface.

This fixes FWNX-881.

Change-Id: Icbc0d3851e1632061e91401d849efd9420a8c8da