Patches for msbuild/xbuild related bugs.
The first patch is a backport of fixes/changes that have already been
implemented in the current development branch of mono.  I found and
reported one of the bugs as Xamarin 6708, and the mono team came up
with a fix after I provided them with a sample set of projects that
demonstrated the bug.  I found the other fixes in the current source
tree after encountering the related problems in our (relatively old)
version of mono.

The second patch is a fix for Xamarin 7594 which I reported, but which
has not yet acted on by the mono team.