Compile libgdiplus with pango
This improves the support for non-roman scripts in the UI. This
(mostly) fixes BL-485.

In order to support non-roman scripts in mono we'll have to use
pango. The non-pango code (which is in src/text-cairo.c in libgdiplus)
uses the "toy" API method cairo_show_text() which doesn't deal with
non-roman nor with having several scripts in one string. Using pango
instead provides all that.

Note that the pango support in libgdiplus is marked as experimental
and unsupported; however with our recent changes in libgdiplus
it seems to work fairly well.

Change-Id: I927e7cee4daca3f02040a16b1265a2cc0b53d056