Create new upstream version numbers
This is necessary to create and upload a source package (the first
version that has a different upstream version number is expected to
create and upload the source package. Future versions with the same
upstream version number won't upload the source package anymore.
Uploading the binary package will fail if the source package can't
be found).

Change-Id: I6ec96c8dda4b65e9e217037c32a224e341ce059f